Brought up another bunch of functions
